Adding the base tests
