+++ monthly/.env
@@ -5,7 +5,9 @@
 # See the documentation for all the connection string options: https://pris.ly/d/connection-strings
 
 DATABASE_URL="mysql://monthly:monthly@localhost:3306/monthly?schema=public"
-DATABASE_TESTING_URL="mysql://monthly:monthly@localhost:3306/testing_monthly?schema=public"
+
+# Uncomment this one out upon testing.
+#DATABASE_URL="mysql://root:root@localhost:3306/testing_monthly?schema=public"
 
 upload_path=/Users/roysegall/Sites/localhost/open_pension/monthly/uploaded_files
