+++ monthly/package.json
@@ -16,7 +16,7 @@
     "worker": "node build/main/job.js",
     "queue:dev": "nodemon --watch '**/*.ts' --exec 'ts-node' src/lib/queue.ts",
     "prisma:makeMigration": "prisma migrate dev",
-    "prisma:migrate": "prisma db push --preview-feature",
+    "prisma:migrate": "prisma db push",
     "prisma:clientGenerate": "prisma generate",
     "prisma:seed": "prisma db seed --preview-feature",
     "prisma:studio": "prisma studio",
@@ -62,6 +62,7 @@
     "@prisma/client": "^2.24.1",
     "@types/xml2js": "^0.4.7",
     "apollo-server": "^2.19.2",
+    "apollo-server-testing": "^2.25.2",
     "bree": "^6.2.0",
     "date-fns": "^2.22.1",
     "dotenv": "^8.2.0",
