+++ monthly/src/server/tests/channel.test.ts
@@ -1,6 +1,19 @@
+// import {prisma} from '../context';
+
+import {createTestingServer, sendQuery} from "./testingServer";
+import {channelsQuery} from "./testingQueries";
+
 describe('Channel', () => {
-  it('Send query when there no channels', () => {
-    expect(1).toBe(1);
+
+  let testingServer;
+
+  beforeAll(() => {
+    testingServer = createTestingServer();
+  });
+
+  it('Send query when there no channels', async () => {
+    const {data: {channels: channels}} = await sendQuery(channelsQuery, testingServer);
+    expect(channels).toStrictEqual([]);
   });
 
   it('Send query when there is a single channel', () => {
