+++ monthly/src/server/tests/testingQueries.ts
@@ -1,0 +1,10 @@
+import {gql} from "apollo-server";
+
+export const channelsQuery = gql`
+  query {
+    channels {
+      ID
+      label
+    }
+  }
+`;
