+++ monthly/src/tests/jestSetup.ts
@@ -1,14 +1,12 @@
 import {prisma} from "../server/context";
 
-let DATABASE_URL;
-
 beforeEach(async () => {
-  DATABASE_URL = process.env.DATABASE_URL;
-  process.env.DATABASE_URL = process.env.DATABASE_TESTING_URL;
   await prisma.$connect();
+
+  const tablesToDelete = [prisma.channel.deleteMany()]
+  await prisma.$transaction(tablesToDelete)
 });
 
 afterAll(async () => {
   await prisma.$disconnect();
-  process.env.DATABASE_URL = DATABASE_URL;
 });
